Support for passing token over cookie
A cookie can be used as transport mechanism for a token.

Name of the cookie is configured through the 'accessTokenName' configuration
option. (Same as query parameter).

The feature can be disabled by flipping the 'allowQueryToken' configuration
option to false. It's enabled by default

Order of precedence for tokens delivery mechanism is:

 1. Authorization header
 2. Cookie
 3. Query parameter